Rename paths and arrange folder structures
