+++ App.js
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import NavStacks from "./Navigation/NavStacks";
+import NavStacks from "./src/navigation/NavStacks";
 import { TailwindProvider } from "tailwindcss-react-native";
 import store from "./features/store";
 import { Provider } from "react-redux";
